graphql_project/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
packaging==25.0
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
